requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
google-generativeai>=0.3.0
markdown>=3.5.0
